Dockerfile.web changed with the new Ruby version and a Bundler update to prevent an error with gems. Several dev gems added. ResetDocker script changed to not destroy images unless said otherwise. Webpack config and assets load fixed. Redis load fixed with its own Dockerfile and script.

+++ scripts/reset-docker.sh
@@ -1,6 +1,16 @@
 #!/bin/bash
 
 docker volume rm $(docker volume ls) # Deletes all the volumes.
-docker rmi -f $(docker images -q)    # Deletes all the images.
 docker stop $(docker ps -aq)         # Stops all the containers.
 docker rm $(docker ps -aq)           # Deletes all the containers.
+
+if [ $# -eq 0 ] ; then
+  echo 'No arguments supplied. Skipping images destroy...'
+else
+  if [ $1 == "--destroy-images" ] ; then
+    echo 'Destroying all images...'
+    docker rmi -f $(docker images -q)  # Deletes all the images.
+  else
+    echo 'Argument supplied different than "--destroy-images". Skipping it...'
+  fi
+fi
